Update structure to scope utils more
